Save before disassembler merge

+++ src/prints.c
@@ -39,7 +39,7 @@ void print_regs_header() {
   printf(" AX   BX   CX   DX   SP   BP   SI   DI  FLAGS IP\n");
 }
 
-void print_regs_status(size_t *r, char *f, char *l, char *m) {
+void print_regs_status(uint16_t *r, char *f, char *l, char *m) {
   printf(REGS_STATUS, r[0], r[1], r[2], r[3], r[4], r[5], r[6], r[7]);
   printf(FLAGS_STATUS, f[0], f[1], f[2], f[3]);
   printf(LINE_STATUS, l);
